Initial commit of TUI using Terminal.Gui 2.0
